# docs/openapi.json
{"components":{"schemas":{"BatteryTestRequest":{"description":"BatteryTestRequest schema","properties":{"action":{"type":"string"}},"required":["action"],"type":"object"},"BeeperRequest":{"description":"BeeperRequest schema","properties":{"enable":{"type":"boolean"}},"type":"object"},"FullStatus":{"description":"FullStatus schema","properties":{"beeper_code":{"type":"integer"},"properties":{},"status":{}},"type":"object"},"HTTPError":{"description":"HTTPError schema","properties":{"detail":{"description":"Human readable error message","nullable":true,"type":"string"},"errors":{"items":{"nullable":true,"properties":{"more":{"additionalProperties":{"description":"Additional information about the error","nullable":true},"description":"Additional information about the error","nullable":true,"type":"object"},"name":{"description":"For example, name of the parameter that caused the error","type":"string"},"reason":{"description":"Human readable error message","type":"string"}},"type":"object"},"nullable":true,"type":"array"},"instance":{"nullable":true,"type":"string"},"status":{"description":"HTTP status code","example":403,"nullable":true,"type":"integer"},"title":{"description":"Short title of the error","nullable":true,"type":"string"},"type":{"description":"URL of the error type. Can be used to lookup the error in a documentation","nullable":true,"type":"string"}},"type":"object"},"MessageResponse":{"description":"MessageResponse schema","properties":{"message":{"type":"string"}},"type":"object"},"Properties":{"description":"Properties schema"},"unknown-interface":{"description":"unknown-interface schema"}}},"info":{"description":"Control and Monitor CyberPower UPS devices via USB HID","title":"CyberPower UPS Monitor","version":"0.0.1"},"openapi":"3.1.0","paths":{"/metrics":{"get":{"description":"#### Controller: \n\n`main.main.func1`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n","operationId":"GET_/metrics","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/unknown-interface"}},"application/xml":{"schema":{"$ref":"#/components/schemas/unknown-interface"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Bad Request _(validation or deserialization error)_"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Internal Server Error _(panics)_"},"default":{"description":""}},"summary":"func1"}},"/ups":{"get":{"description":"#### Controller: \n\n`main.listDevices`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n","operationId":"GET_/ups","parameters":[{"in":"header","name":"Accept","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Properties"},"type":"array"}},"application/xml":{"schema":{"items":{"$ref":"#/components/schemas/Properties"},"type":"array"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Bad Request _(validation or deserialization error)_"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Internal Server Error _(panics)_"},"default":{"description":""}},"summary":"list devices"}},"/ups/{serial}":{"get":{"description":"#### Controller: \n\n`main.getDeviceStatus`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n","operationId":"GET_/ups/:serial","parameters":[{"in":"header","name":"Accept","schema":{"type":"string"}},{"in":"path","name":"serial","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FullStatus"}},"application/xml":{"schema":{"$ref":"#/components/schemas/FullStatus"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Bad Request _(validation or deserialization error)_"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Internal Server Error _(panics)_"},"default":{"description":""}},"summary":"get device status"}},"/ups/{serial}/battery-test":{"post":{"description":"#### Controller: \n\n`main.runBatteryTest`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n","operationId":"POST_/ups/:serial/battery-test","parameters":[{"in":"header","name":"Accept","schema":{"type":"string"}},{"in":"path","name":"serial","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/BatteryTestRequest"}}},"description":"Request body for main.BatteryTestRequest","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}},"application/xml":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Bad Request _(validation or deserialization error)_"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Internal Server Error _(panics)_"},"default":{"description":""}},"summary":"run battery test"}},"/ups/{serial}/beeper":{"post":{"description":"#### Controller: \n\n`main.setBeeper`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n","operationId":"POST_/ups/:serial/beeper","parameters":[{"in":"header","name":"Accept","schema":{"type":"string"}},{"in":"path","name":"serial","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/BeeperRequest"}}},"description":"Request body for main.BeeperRequest","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}},"application/xml":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Bad Request _(validation or deserialization error)_"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPError"}},"application/xml":{"schema":{"$ref":"#/components/schemas/HTTPError"}}},"description":"Internal Server Error _(panics)_"},"default":{"description":""}},"summary":"set beeper"}}},"servers":[{"url":"http://localhost:9999"},{"description":"local server","url":"http://:9999"}]}